GCC compability; removed ::juce from global namespace due to incompabilites with X11 headers

+++ gui/controls/CComboBox.cpp
@@ -1,30 +1,30 @@
 /*************************************************************************************
- 
+
 	cpl - cross-platform library - v. 0.1.0.
- 
+
 	Copyright (C) 2016 Janus Lynggaard Thorborg (www.jthorborg.com)
- 
+
 	This program is free software: you can redistribute it and/or modify
 	it under the terms of the GNU General Public License as published by
 	the Free Software Foundation, either version 3 of the License, or
 	(at your option) any later version.
- 
+
 	This program is distributed in the hope that it will be useful,
 	but WITHOUT ANY WARRANTY; without even the implied warranty of
 	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 	GNU General Public License for more details.
- 
+
 	You should have received a copy of the GNU General Public License
 	along with this program.  If not, see <http://www.gnu.org/licenses/>.
- 
+
 	See \licenses\ for additional details on licenses associated with this program.
- 
+
 **************************************************************************************
- 
+
 	file:CComboBox.cpp
- 
+
 		Source code for CComboBox.h
- 
+
 *************************************************************************************/
 
 #include "CComboBox.h"
@@ -101,12 +101,12 @@ namespace cpl
 		auto len = inputValues.size();
 		for (std::size_t i = 0; i < len; i++)
 		{
-			if (inputValues[i] == '|') 
+			if (inputValues[i] == '|')
 			{
 				values.push_back(inputValues.substr(iter_pos, i - iter_pos));
 				iter_pos = i + 1;
 			}
-			else if (i == (len - 1)) 
+			else if (i == (len - 1))
 			{
 				values.push_back(inputValues.substr(iter_pos, i + 1 - iter_pos));
 
@@ -154,7 +154,7 @@ namespace cpl
 		if (2 > size)
 			size = 2;
 		idx = cpl::Math::confineTo(idx, 1, size);
-		
+
 		return iCtrlPrec_t(idx - 1) / (size - 1);
 	}
 
@@ -188,7 +188,7 @@ namespace cpl
 
 	void CComboBox::bSetValue(iCtrlPrec_t val, bool sync)
 	{
-		box.setSelectedId(floatToInt(val, (int)values.size()), 
+		box.setSelectedId(floatToInt(val, (int)values.size()),
 			sync ? juce::NotificationType::sendNotificationSync : juce::NotificationType::sendNotification);
 	}
 
@@ -224,7 +224,7 @@ namespace cpl
 	bool CComboBox::bStringToValue(const std::string & valueString, iCtrlPrec_t & val) const
 	{
 		auto idx = indexOfValue(valueString);
-		if (idx != -1)
+		if (idx != static_cast<std::size_t>(-1))
 		{
 			// can never divide by zero here, since the for loop won't enter in that case
 			//val = iCtrlPrec_t(i) / (values.size() > 1 ? values.size() - 1 : 1);
@@ -254,7 +254,7 @@ namespace cpl
 
 	bool CComboBox::setEnabledStateFor(std::size_t entry, bool toggle)
 	{
-		if (entry < values.size() && entry != -1)
+		if (entry < values.size() && entry != static_cast<std::size_t>(-1))
 		{
 			bool selectNew = box.getSelectedId() == int(entry + 1) && !toggle;
 			box.setItemEnabled((int)entry + 1, !!toggle);
